+++ packages/ast-manipulation/js/common.ts
@@ -1,5 +1,6 @@
 import { type AstKinds, type AstTypes, Walker, parseScript, serializeScript, stripAst } from "@svelte-add/ast-tooling";
 import decircular from "decircular";
+import dedent from "dedent";
 
 export function addJsDocTypeComment(node: AstTypes.Node, type: string) {
     const comment: AstTypes.CommentBlock = {
@@ -20,6 +21,13 @@ export function typeAnnotateExpression(node: AstKinds.ExpressionKind, type: stri
     };
 
     return expression;
+}
+
+export function createSpreadElement(expression: AstKinds.ExpressionKind): AstTypes.SpreadElement {
+    return {
+        type: "SpreadElement",
+        argument: expression,
+    };
 }
 
 export function createLiteral(value: string | null = null) {
@@ -58,7 +66,7 @@ export function expressionStatement(expression: AstKinds.ExpressionKind) {
 }
 
 export function addFromString(ast: AstTypes.BlockStatement | AstTypes.Program, value: string) {
-    const program = parseScript(value);
+    const program = parseScript(dedent(value));
 
     for (const childNode of program.body) {
         ast.body.push(childNode);
@@ -66,7 +74,7 @@ export function addFromString(ast: AstTypes.BlockStatement | AstTypes.Program, v
 }
 
 export function expressionFromString(value: string): AstKinds.ExpressionKind {
-    const program = parseScript(value);
+    const program = parseScript(dedent(value));
     const statement = program.body[0];
     if (statement.type !== "ExpressionStatement") {
         throw new Error("value passed was not an expression");
@@ -76,7 +84,7 @@ export function expressionFromString(value: string): AstKinds.ExpressionKind {
 }
 
 export function statementFromString(value: string): AstKinds.StatementKind {
-    const program = parseScript(value);
+    const program = parseScript(dedent(value));
     const statement = program.body[0];
 
     return statement;